requirements/base.txt
Django==4.0.7
django-environ==0.9.0
psycopg2-binary==2.9.5
djangorestframework==3.13.1

django-filter==22.1
django-extensions==3.2.1
django-cors-headers==3.13.0
whitenoise==6.2.0
drf-jwt==1.19.2

djangorestframework-simplejwt==5.2.2
drf-spectacular==0.24.2

pytest==7.2.0
pytest-django==4.5.2
